Analysis of Line x Tester data laid out in Alpha Lattice experimental design

Used for analyzing the line x tester data containing only crosses laid out in alpha lattice design

ltcalpha(data, block, replication, line, tester, y)

Arguments

data
dataframe containing following variables

block
number of blocks per replication

replication
number of replications

line
number of lines

tester
number of testers

y
a trait of interest

Value

Results of line by tester analysis

Details

Analyzing the line by tester data only using the data from crosses which are evaluated in alpha lattice design.

Author

Nandan Patil

Examples

library(gpbStat)
data(alphalt)
results1 = ltcalpha(alphalt, replication, block, line, tester, yield)

#> 
#> Analysis of Line x Tester:  yield 

results1

#> $Means
#>      Testers
#> Lines        6        7        8
#>     1 86.47500 88.95833 89.55000
#>     2 88.64667 55.48000 50.12667
#>     3 51.19917 53.28417 36.91583
#>     4 33.47500 34.29833 50.78417
#>     5 45.30417 42.14500 49.98000
#> 
#> $ANOVA.of.Crosses
#>                           Df    Sum Sq   Mean Sq   F value       Pr(>F)
#> Replication                3  1586.493  528.8311  3.566866 0.0549612862
#> Crosses                   14 23862.020 1704.4300 11.496059 0.0002347628
#> Blocks within Replication 32  5446.515  170.2036  1.147991 0.4316073578
#> Error                     10  1482.621  148.2621        NA           NA
#> 
#> $Coefficient.of.Variation
#> [1] 21.32146
#> 
#> $ANOVA.for.Line.by.Tester.analysis
#>                 Df     Sum Sq   Mean Sq   F value       Pr(>F)
#> Lines            4 18835.3119 4708.8280 24.883334 6.536498e-11
#> Testers          2   463.1458  231.5729  1.223724 3.037332e-01
#> Lines X Testers  8  4563.5622  570.4453  3.014461 8.508293e-03
#> Error           10  1482.6212  148.2621        NA           NA
#> 
#> $GCA.lines
#>       1       2       3       4       5 
#>  31.220   7.643  -9.975 -17.589 -11.298 
#> 
#> $GCA.testers
#>      6      7      8 
#>  3.912 -2.275 -1.637 
#> 
#> $Proportional.Contribution.of.lines.testers.and.line.by.tester.interaction.to.total.variance
#>          Lines         Tester  Line x Tester 
#>      78.934273       1.940933      19.124794 
#> 
#> $Genetic.Components.of.Variance
#>                  Cov H.S. (line)                Cov H.S. (tester) 
#>                        344.86523                        -16.94362 
#>               Cov H.S. (average)               Cov F.S. (average) 
#>                         30.06778                        263.22720 
#> F = 0, Adittive genetic variance F = 1, Adittive genetic variance 
#>                        120.27111                         60.13555 
#> F = 0, Variance due to Dominance F = 1, Variance due to Dominance 
#>                        211.09158                         20.82769 
#> 
#> $Standard.Errors
#>      S.E. gca for line    S.E. gca for tester        S.E. sca effect 
#>               3.514993               2.722702               6.088147 
#>     S.E. (gi - gj)line   S.E. (gi - gj)tester S.E. (sij - skl)tester 
#>               4.970951               3.850482               8.609940 
#> 
#> $Critical.differance.at.5.percent
#>      C.D. gca for line    C.D. gca for tester        C.D. sca effect 
#>               7.831893               6.066558              13.565236 
#>     C.D. (gi - gj)line   C.D. (gi - gj)tester C.D. (sij - skl)tester 
#>              11.075969               8.579409              19.184141 
#>