Analysis of Line x Tester data containing crosses and checks laid out in RCBD or Alpha Lattice experimental design

Analysis of Line x Tester data containing crosses and checks laid out in RCBD or Alpha Lattice experimental design

ltcchk(data, replication, line, tester, check, y, block)

Arguments

data
dataframe containing following variables

replication
replication variable

line
line variable

tester
tester variable

check
check variable

y
trait of interest

block
block variable (for alpha lattice only)

Value

Overall ANOVA

ANOVA with all the factors.

Coefficient of Variation

ANOVA with all the factors.

Genetic Variance

Phenotypic and Genotypic variance for the given trait.

Genetic Variability

Phenotypic coefficient of variability and Genotypic coefficient of variability and Environmental coefficient of Variation.

Proportional Contribution

Propotional contribution of Lines, Tester and Line x Tester interaction.

GCA lines

Combining ability effects of lines.

GCA testers

Combining ability effects of testers.

SCA crosses

Combining ability effects of crosses

Line x Tester ANOVA

ANOVA with all the factors.

GV Singh & Chaudhary

Genetic component of Variance as per Singh and Chaudhary, 1977.

GV King

Genetic component of Variance as per King, 1961.

Standard Errors

Standard error for combining ability effects.

Critical Difference

Critical Difference at 5 percent for combining ability effects.

Details

Analyzing the line by tester data only using the data from crosses which are evaluated in alpha lattice design. All the factors are considered as fixed.

Analyzing the line by tester data only using the data from crosses which are evaluated in alpha lattice design. All the factors are considered as fixed.

Note

The block variable is inserted at the last if the experimental design is Alpha Lattice. For RCBD no need to have block factor.

References

Kempthorne, O. (1957), Introduction to Genetic Statistics. John Wiley and Sons, New York. , 468-472. Singh, R. K. and Chaudhary, B. D. (1977). Biometrical Methods in Quantitative Genetic Analysis. Kalyani Publishers, New Delhi. King, J. G., Quinby, J. R., Stephens, j. C., Kramer, N. W. and Lahr, K. A. (1961). An evaluation of parents of grain Sorghum hybrids. Bull. Texas Agric. Exp. Sta., M.P. 510.

See also

Author

Nandan Patil

Nandan Patil patilnads@gmail.com

Examples