Analysis of Line x Tester data laid out in RCBD experimental design

Used for analyzing the line x tester data containing only crosses laid out in RCBD design

ltcrcbd(data, replication, line, tester, y)

Arguments

data
dataframe containing following variables

replication
number of replications

line
number of lines

tester
number of testers

y
trait of interest

Value

Results of line by tester analysis

Details

Analyzing the line by tester data containing only crosses which are evaluated in Randomized Complete Block Design (RCBD).

Author

Nandan Patil

Examples

library(gpbStat)
data(rcbdlt)
results2 = ltcrcbd(rcbdlt, replication, line, tester, yield)

#> 
#> Analysis of Line x Tester:  yield 

results2

#> $Means
#>      Testers
#> Lines       6       7       8
#>     1  68.550 107.640  52.640
#>     2  73.265  97.640  85.650
#>     3 100.885 111.540 117.735
#>     4 105.795  64.450  46.855
#>     5  84.150  81.935  94.820
#> 
#> $ANOVA.of.Crosses
#>             Df    Sum Sq    Mean Sq   F value       Pr(>F)
#> Replication  3   148.436   49.47866  0.509612 6.778194e-01
#> Crosses     14 26199.654 1871.40388 19.274772 6.737492e-14
#> Error       42  4077.815   97.09084        NA           NA
#> Total       59 30425.906         NA        NA           NA
#> 
#> $Coefficient.of.Variation
#> [1] 11.42608
#> 
#> $ANOVA.for.Line.by.Tester.analysis
#>                 Df    Sum Sq    Mean Sq   F value       Pr(>F)
#> Lines            4 10318.361 2579.59035 27.466791 1.421271e-11
#> Testers          2  1718.926  859.46289  9.151332 4.626865e-04
#> Lines X Testers  8 14162.367 1770.29589 18.849639 4.973396e-12
#> Error           42  4077.815   97.09084        NA           NA
#> 
#> $GCA.lines
#>       1       2       3       4       5 
#>  -9.960  -0.718  23.817 -13.870   0.732 
#> 
#> $GCA.testers
#>      6      7      8 
#>  0.292  6.404 -6.697 
#> 
#> $Proportional.Contribution.of.lines.testers.and.line.by.tester.interaction.to.total.variance
#>          Lines         Tester  Line x Tester 
#>      39.383578       6.560872      54.055550 
#> 
#> $Genetic.Components.of.Variance
#>                  Cov H.S. (line)                Cov H.S. (tester) 
#>                        67.441205                       -45.541650 
#>               Cov H.S. (average)               Cov F.S. (average) 
#>                         2.680894                       408.052454 
#> F = 0, Adittive genetic variance F = 1, Adittive genetic variance 
#>                        10.723574                         5.361787 
#> F = 0, Variance due to Dominance F = 1, Variance due to Dominance 
#>                       836.602526                       418.301263 
#> 
#> $Standard.Errors
#>      S.E. gca for line    S.E. gca for tester        S.E. sca effect 
#>               2.844451               2.203303               4.926734 
#>     S.E. (gi - gj)line   S.E. (gi - gj)tester S.E. (sij - skl)tester 
#>               4.022662               3.115940               6.967454 
#> 
#> $Critical.differance.at.5.percent
#>      C.D. gca for line    C.D. gca for tester        C.D. sca effect 
#>               5.740335               4.446445               9.942552 
#>     C.D. (gi - gj)line   C.D. (gi - gj)tester C.D. (sij - skl)tester 
#>               8.118060               6.288222              14.060892 
#>